Renamed quartic solver module names
