+++ src/quartics/ferrari.py
