+++ src/quartics/numpyqs.py
